Replace @turf/turf
While @turf/turf is licensed under the MIT license, two of its
component libraries, `@turf/isolines` and `@turf/isobands`, depend on
`marchingsquares`, which is licensed under the AGPL 3.0. This renders
the full @turf/turf library, and any library that depends on it,
unusable by most organizations. While I understand and support the
goals of the AGPL, it is not an appropriate license for library code.
Any code that depends on AGPL code is tainted and unusable because of
it.

Thankfully, in the case of turf, there is a workaround: the component
libraries are all published seperately as well, so we as long as
there's no dependency on `@turf/isolines` or `@turf/isobands` they
can be installed without the virality of the AGPL being a concern.

This pull request does just that.